requirements.txt
requests
pydantic
braceexpand
appdirs
pyyaml
